Add support for yields crossing ShadowDOM barrier
Prior to this change, the yields event would bubble/propagate through
the LightDOM but the ShadowDOM did not propagate through. This meant
that components that used a template which rendered another component
the child component would not have access to the parent's yields().

This change adds an event listener to this.shadow as well to capture any
components either in the LightDOM or the ShadowDOM.

Fixes #8