Retain slot names in LightDOM implementation
Prior to this change, we removed the slot name attributes but this
de-identified the content from the containing component in the LightDOM.
This meant that the containing component needed implicit knowledge of
the HTML it was given.

This change keeps the slot name in place so the containing component can
have something it knows to reference. (default slot doesn't have
a corresponding identifier thus there isn't much we can/should do in
this case it is on the component makers to wrap default slots in
something identifiable. Closest approximation of the ShadowDOM.

    class FooBar extends Component {
      render() {
        this.querySelectorAll('[slot=foo]').forEach(…);
      }
      static get template() {
        return `<slot name="foo"></slot><slot></slot>`;
      }
    }
    FooBar.register();

    <foo-bar>
      <p slot="foo">
        <code>[slot=foo]</code> should find <em>this</em> P tag.
      </p>
    </foo-bar>