Add tests suite
Had to guard against the yieldableProxy being triggered for non-string
properties. This happened when using SinonJS for inspection of spy
arguments. It used Symbol.toStringTag for inspection which was not the
intent of the yieldableProxy. Others could run into this if they console
logged or other manipulations to the yieldableProxy. This it was best to
only react if the use case was a real property I.E. a string.